Includes ofxOsc properly
Had to rebuild the project with the project generator to include the
addon.

+++ src/Particle.h
@@ -16,7 +16,7 @@ public:
 	void yesGravity();
     void updateGravity();
 	
-	void report(float collision);
+	void report(int collision);
 	
 	int getSpeed() { return speed; }
 	int getGravity() { return gravity; }
